fix: satisfy ShellCheck 0.11

+++ scripts/provision-common.sh
@@ -149,7 +149,7 @@ publish_completion_release() (
 
   case $completion_old_pointer in
     .dotfiles-completions-release.*)
-      rm -rf "$completion_dir/$completion_old_pointer"
+      rm -rf -- "${completion_dir:?}/${completion_old_pointer:?}"
       ;;
   esac
 )
